% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_attribute_averages.R
\name{exist_attribute_averages}
\alias{exist_attribute_averages}
\title{Query Exist.io for all averages for a specific attribute}
\usage{
exist_attribute_averages(
  token = NULL,
  attribute = "mood",
  limit = NULL,
  page = 1,
  date_min = NULL,
  date_max = NULL
)
}
\arguments{
\item{token}{The token environment.}

\item{attribute}{The attribute for which to get all insights. Defaults to mood.}

\item{limit}{Number of values to return per page, starting with today. Optional, max is 100.}

\item{page}{Page index. Optional, default is 1.}

\item{date_min}{Oldest date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.}

\item{date_max}{Most recent date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.}
}
\value{
A list containing the information of all attributes for a specific attribute.
}
\description{
Returns information on all averages for a specific attribute.
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_attribute_averages(etoken)
}
}
