% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_attribute_correlations.R
\name{exist_attribute_correlations}
\alias{exist_attribute_correlations}
\title{Query Exist.io for all correlations for a specific attribute}
\usage{
exist_attribute_correlations(
  token = NULL,
  attribute = "mood",
  limit = NULL,
  page = 1,
  date_min = NULL,
  date_max = NULL,
  latest = NULL
)
}
\arguments{
\item{token}{The token environment.}

\item{attribute}{The attribute for which to get all correlations. Defaults to mood.}

\item{limit}{Number of values to return per page, starting with today. Optional, max is 100.}

\item{page}{Page index. Optional, default is 1.}

\item{date_min}{Oldest date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.}

\item{date_max}{Most recent date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.}

\item{latest}{Set this to true to return only the most recently generated batch of correlations. Use this on its own without date_min and date_max.}
}
\value{
A list containing the information of all correlations for a specific attribute.
}
\description{
Returns information on all correlations for a specific attribute.
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_attribute_correlations(etoken)
}
}
