% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_auth.R
\name{exist_auth}
\alias{exist_auth}
\title{Authenticate to Exist.io}
\usage{
exist_auth()
}
\description{
Authenticates the user to Exist.io using OAuth2.0.
}
\details{
This function stores the users auth token localy. This requires the client id and client secret to be available as environment variables.
}
\examples{
\donttest{
Sys.setenv(EXIST_CLIENT_ID = "PASTE_YOUR_ID_HERE")
Sys.setenv(EXIST_CLIENT_SECRET = "PASTE_YOUR_SECRET_HERE")

exist_auth()
}
}
