% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_current_averages.R
\name{exist_current_averages}
\alias{exist_current_averages}
\title{Query Exist.io for the most recent averages.}
\usage{
exist_current_averages(token = NULL)
}
\arguments{
\item{token}{The token environment.}
}
\value{
A list containing the most recent averages.
}
\description{
Returns the most recent average values for each attribute.
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_current_averages(etoken)
}
}
