% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_multiple_attributes.R
\name{exist_multiple_attributes}
\alias{exist_multiple_attributes}
\title{Query Exist.io for multiple attributes}
\usage{
exist_multiple_attributes(
  token = NULL,
  limit = NULL,
  attributes = NULL,
  groups = NULL,
  date_max = NULL
)
}
\arguments{
\item{token}{The token environment.}

\item{limit}{Number of values to return, starting with the newest date. Optional, max is 31.}

\item{attributes}{Optional comma-separated list of attributes, e.g. mood,mood_note.}

\item{groups}{Optional comma-separated list of groups to filter attributes by, e.g. mood,health}

\item{date_max}{Optional date specifying the newest value that can be returned, in format YYYY-mm-dd.}
}
\value{
A list containing the information about the current day of the user.
}
\description{
Returns information on multiple attributes.
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_multiple_attributes()
}
}
