% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_owned_attributes.R
\name{exist_owned_attributes}
\alias{exist_owned_attributes}
\title{Query Exist.io for the owned attributes}
\usage{
exist_owned_attributes()
}
\value{
A list containing the owned attributes.
}
\description{
Returns information on the owned attributes.
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_owned_attributes()
}
}
