% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_package_query.R
\name{exist_package_query}
\alias{exist_package_query}
\title{Query Exist.io API}
\usage{
exist_package_query(path = "api/1/users/$self/profile/", query = NULL)
}
\arguments{
\item{path}{The API path to query.}

\item{query}{A list of query parameters.}
}
\value{
The response of the API call
}
\description{
Returns information on the logged in user.
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_package_query(path = "api/1/users/$self/profile/")
}
}
