% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_package_useragent.R
\name{exist_package_useragent}
\alias{exist_package_useragent}
\title{existR user agent}
\usage{
exist_package_useragent()
}
\value{
A character string containing the user agent used in API calls generate with the existR package.
}
\description{
This function returns the user agent used in API calls of the existR package.
}
\examples{
exist_package_useragent()
}
