% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_release_attribute.R
\name{exist_release_attribute}
\alias{exist_release_attribute}
\title{Release ownership for an attribute on Exist.io}
\usage{
exist_release_attribute(attribute = NULL)
}
\arguments{
\item{attribute}{The attribute for which to release ownership.}
}
\value{
A status indicating if the request was successful.
}
\description{
Release ownership for an attribute on Exist.io
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_release_attribute()
}
}
