% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_strongest_correlations.R
\name{exist_strongest_correlations}
\alias{exist_strongest_correlations}
\title{Query Exist.io for the strongest correlations across all attributes.}
\usage{
exist_strongest_correlations()
}
\value{
A list containing the strongest correlations across all attributes.
}
\description{
Returns a list of the user’s strongest correlations across all attributes.
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_strongest_correlations()
}
}
