% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_today.R
\name{exist_today}
\alias{exist_today}
\title{Query Exist.io current users day}
\usage{
exist_today(token = NULL)
}
\arguments{
\item{token}{The token environment.}
}
\value{
A list containing the information about the current day of the user.
}
\description{
Returns information on the current day of the logged in user.
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_today(etoken)
}
}
