% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_update_attribute.R
\name{exist_update_attribute}
\alias{exist_update_attribute}
\title{Update an attribute on Exist.io}
\usage{
exist_update_attribute(attribute = NULL, date = NULL, value = NULL)
}
\arguments{
\item{attribute}{The attribute to update.}

\item{date}{String of format YYYY-mm-dd.}

\item{value}{A valid value for this attribute type: string, integer, or float.}
}
\value{
A status indicating if the request was successful.
}
\description{
Update an attribute on Exist.io
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_update_attribute()
}
}
