% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_all_insights.R
\name{print.exist_all_insights}
\alias{print.exist_all_insights}
\title{Print result of exist_all_insights()}
\usage{
\method{print}{exist_all_insights}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the exist_all_insights() function.
}
\description{
Print a structured return of the exist_all_insights() function.
}
\examples{
\dontrun{
resp <- exist_all_insights()

resp
}
}
\seealso{
\code{\link{exist_all_insights}}
}
