% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_append_tag.R
\name{print.exist_append_tag}
\alias{print.exist_append_tag}
\title{Print result of exist_append_tag()}
\usage{
\method{print}{exist_append_tag}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the exist_append_tag() function.
}
\description{
Print a structured return of the exist_append_tag() function.
}
\examples{
\dontrun{
resp <- exist_append_tag()

resp
}
}
\seealso{
\code{\link{exist_append_tag}}
}
