% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_attribute_averages.R
\name{print.exist_attribute_averages}
\alias{print.exist_attribute_averages}
\title{Print result of exist_attribute_averages()}
\usage{
\method{print}{exist_attribute_averages}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the exist_attribute_averages() function.
}
\description{
Print a structured return of the exist_attribute_averages() function.
}
\examples{
\dontrun{
resp <- exist_attribute_averages()

resp
}
}
\seealso{
\code{\link{exist_attribute_averages}}
}
