% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_multiple_attributes.R
\name{print.exist_multiple_attributes}
\alias{print.exist_multiple_attributes}
\title{Print result of exist_multiple_attributes()}
\usage{
\method{print}{exist_multiple_attributes}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the exist_multiple_attributes() function.
}
\description{
Print a structured return of the exist_multiple_attributes() function.
}
\examples{
\dontrun{
resp <- exist_multiple_attributes()

resp
}
}
\seealso{
\code{\link{exist_multiple_attributes}}
}
