% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_profile.R
\name{print.exist_profile}
\alias{print.exist_profile}
\title{Print result of exist_profile()}
\usage{
\method{print}{exist_profile}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the exist_profile() function.
}
\description{
Print a structured return of the exist_profile() function.
}
\examples{
\donttest{
resp <- exist_profile()

resp
}
}
\seealso{
\code{\link{exist_profile}}
}
