% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_strongest_correlations.R
\name{print.exist_strongest_correlations}
\alias{print.exist_strongest_correlations}
\title{Print result of exist_strongest_correlations()}
\usage{
\method{print}{exist_strongest_correlations}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the exist_strongest_correlations() function.
}
\description{
Print a structured return of the exist_strongest_correlations() function.
}
\examples{
\dontrun{
resp <- exist_strongest_correlations()

resp
}
}
\seealso{
\code{\link{exist_strongest_correlations}}
}
