% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_today.R
\name{print.exist_today}
\alias{print.exist_today}
\title{Print result of exist_today()}
\usage{
\method{print}{exist_today}(x, ...)
}
\arguments{
\item{x}{The return of the function this print function relates to.}

\item{...}{Possible further options to the print function.}
}
\value{
A structured print of the return by the exist_today() function.
}
\description{
Print a structured return of the exist_today() function.
}
\examples{
\donttest{
resp <- exist_today()

resp
}
}
\seealso{
\code{\link{exist_today}}
}
