% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_acquire_attribute.R
\name{exist_acquire_attribute}
\alias{exist_acquire_attribute}
\title{Acquire ownership for an attribute on Exist.io}
\usage{
exist_acquire_attribute(attribute = NULL, active = "true")
}
\arguments{
\item{attribute}{The attribute for which to acquire ownership.}

\item{active}{true or false to set this attribute to active or inactive. Defaults to true.}
}
\value{
A status indicating if the request was successful.
}
\description{
Acquire ownership for an attribute on Exist.io
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_acquire_attribute()
}
}
