% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_all_insights.R
\name{exist_all_insights}
\alias{exist_all_insights}
\title{Query Exist.io for all insights}
\usage{
exist_all_insights(
  token = NULL,
  limit = NULL,
  page = 1,
  date_min = NULL,
  date_max = NULL
)
}
\arguments{
\item{token}{The token environment.}

\item{limit}{Number of values to return per page, starting with today. Optional, max is 100.}

\item{page}{Page index. Optional, default is 1.}

\item{date_min}{Oldest date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.}

\item{date_max}{Most recent date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.}
}
\value{
A list containing the information of all insights.
}
\description{
Returns information on all insights.
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_all_insights(etoken)
}
}
