% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_append_tag.R
\name{exist_append_tag}
\alias{exist_append_tag}
\title{Append a tag on Exist.io}
\usage{
exist_append_tag(value = NULL, date = NULL)
}
\arguments{
\item{value}{The tag to append.}

\item{date}{Optional string of format YYYY-mm-dd. If omitted, current day is assumed.}
}
\value{
A status indicating if the request was successful.
}
\description{
Append a tag on Exist.io
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_append_tag()
}
}
