% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_current_averages.R
\name{exist_current_averages}
\alias{exist_current_averages}
\title{Query Exist.io for the most recent averages.}
\usage{
exist_current_averages()
}
\value{
A list containing the most recent averages.
}
\description{
Returns the most recent average values for each attribute.
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_current_averages()
}
}
