% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_profile.R
\name{exist_profile}
\alias{exist_profile}
\title{Query Exist.io current user}
\usage{
exist_profile()
}
\value{
A list containing the information about the current user.
}
\description{
Returns information on the logged in user.
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_profile()
}
}
