% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_release_attribute.R
\name{exist_release_attribute}
\alias{exist_release_attribute}
\title{Release ownership for an attribute on Exist.io}
\usage{
exist_release_attribute(token = NULL, attribute = NULL)
}
\arguments{
\item{token}{The token environment.}

\item{attribute}{The attribute for which to release ownership.}
}
\value{
A status indicating if the request was successful.
}
\description{
Release ownership for an attribute on Exist.io
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_release_attribute(etoken)
}
}
