% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_specific_attribute.R
\name{exist_specific_attribute}
\alias{exist_specific_attribute}
\title{Query Exist.io for a specific attribute}
\usage{
exist_specific_attribute(
  attribute = "",
  limit = NULL,
  page = 1,
  date_max = NULL
)
}
\arguments{
\item{attribute}{The name of the attribute to get.}

\item{limit}{Number of values to return per page. Optional, max is 100.}

\item{page}{Page index. Optional, default is 1.}

\item{date_max}{Most recent date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.}
}
\value{
A list containing the information about the current day of the user.
}
\description{
Returns information on a specific attribute.
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_specific_attribute()
}
}
