% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_strongest_correlations.R
\name{exist_strongest_correlations}
\alias{exist_strongest_correlations}
\title{Query Exist.io for the strongest correlations across all attributes.}
\usage{
exist_strongest_correlations(token = NULL)
}
\arguments{
\item{token}{The token environment.}
}
\value{
A list containing the strongest correlations across all attributes.
}
\description{
Returns a list of the user’s strongest correlations across all attributes.
}
\examples{
\dontrun{
etoken <- exist_auth()
exist_strongest_correlations(etoken)
}
}
