% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_today.R
\name{exist_today}
\alias{exist_today}
\title{Query Exist.io current users day}
\usage{
exist_today()
}
\value{
A list containing the information about the current day of the user.
}
\description{
Returns information on the current day of the logged in user.
}
\details{
This function expects the user to have the oauth token cached.
}
\examples{
\dontrun{
exist_today()
}
}
